pyproject.toml
[project]
name = "extendicare-kg-rag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "llama-cloud-services>=0.6.30",
    "llama-index-llms-openai>=0.4.3",
    "llama-parse>=0.6.30",
    "plotly>=6.1.2",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
